+++ comics/comics/abstrusegoose.py
@@ -10,6 +10,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 10
+    schedule = 'Mo,Th'
 
     def crawl(self, pub_date):
         feed = self.parse_feed('http://abstrusegoose.com/feed/atom')