+++ comics/comics/basicinstructions.py
@@ -10,7 +10,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 100
-    schedule = 'Mo,We,Fr'
+    schedule = 'Tu,Th,Su'
     time_zone = -7
 
     def crawl(self, pub_date):