+++ comics/comics/bgobt.py
@@ -6,19 +6,10 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.businessguysonbusinesstrips.com/'
     start_date = '2007-07-12'
+    end_date = '2011-11-23'
+    active = False
     rights = '"Managing Director"'
 
 class Crawler(CrawlerBase):
-    history_capable_days = 100
-    time_zone = -7
-
     def crawl(self, pub_date):
-        feed = self.parse_feed(
-            'http://businessguysonbusinesstrips.com/?feed=atom')
-        for entry in feed.for_date(pub_date):
-            page = self.parse_page(entry.link)
-            page.remove('img[src$="/art/wgp_banner.jpg"]')
-            url = page.src(
-                'img[src^="http://businessguysonbusinesstrips.com/art/"]')
-            title = entry.title
-            return CrawlerImage(url, title)
+        pass # Comic no longer published