+++ comics/comics/darklegacy.py
@@ -10,7 +10,6 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_date = '2006-12-09'
-    schedule = 'Su'
     time_zone = -6
 
     def crawl(self, pub_date):