+++ comics/comics/littlegamers.py
@@ -10,7 +10,6 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_date = '2000-12-01'
-    schedule = 'Mo,We'
     time_zone = 1
 
     def crawl(self, pub_date):