+++ comics/comics/picturesforsadchildren.py
@@ -9,13 +9,11 @@ class ComicData(ComicDataBase):
     rights = 'John Campbell'
 
 class Crawler(CrawlerBase):
-    history_capable_days = 40
-    schedule = 'Mo'
+    history_capable_days = 90
     time_zone = -6
 
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://www.rsspect.com/rss/pfsc.xml')
+        feed = self.parse_feed('http://picturesforsadchildren.com/rss')
         for entry in feed.for_date(pub_date):
-            url = entry.summary.src('img[src*="/comics/"]')
-            title = entry.title
-            return CrawlerImage(url, title)
+            url = entry.summary.src('img')
+            return CrawlerImage(url)