+++ comics/comics/radiogaga.py
@@ -10,7 +10,7 @@ class ComicData(ComicDataBase):
     rights = 'Øyvind Sagåsen'
 
 class Crawler(PondusNoCrawlerBase):
-    schedule = 'Mo,Tu,We,Fr'
+    schedule = 'Mo,Tu,We,Th,Fr'
 
     def crawl(self, pub_date):
         return self.crawl_helper('20057')