+++ comics/comics/tehgladiators.py
@@ -10,7 +10,6 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 90
-    schedule = 'Su'
     time_zone = 1
 
     def crawl(self, pub_date):