+++ comics/comics/uvod.py
@@ -4,7 +4,7 @@ from comics.core.comic_data import ComicDataBase
 class ComicData(ComicDataBase):
     name = 'The Unspeakable Vault (of Doom)'
     language = 'en'
-    url = 'http://www.macguff.fr/goomi/unspeakable/'
+    url = 'http://www.goominet.com/unspeakable-vault/'
     rights = 'Francois Launet'
 
 class Crawler(CrawlerBase):