+++ comics/comics/virtualshackles.py
@@ -12,7 +12,6 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 32
-    schedule = 'We,Fr'
 
     def crawl(self, pub_date):
         feed = self.parse_feed('http://feeds.feedburner.com/VirtualShackles')